backend/pyproject.toml
[project]
name = "m-clone-backend"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic-ai[openai]>=0.0.20",
    "pydantic-settings>=2.0",
    "pdfplumber>=0.11",
    "python-multipart>=0.0.9",
    "yfinance>=0.2.40",
    "Wikipedia-API>=0.7",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "tavily-python>=0.5",
    "asyncpg>=0.29",
    "rank-bm25>=0.2",
    "redis[asyncio]>=5.0",
    "authlib>=1.3",
    "python-jose[cryptography]>=3.3",
    "itsdangerous>=2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
